[#10] Remove validation check from vm
